requirements.txt
Flask
Flask-Login
Flask-PyMongo
python-dotenv
werkzeug
Flask-Session
bson
